requirements.txt
pyspark==3.4.0
boto3==1.26.0
snowflake-connector-python<3.0.0
snowflake-sqlalchemy==1.4.0
kafka-python==2.0.2
configparser==5.3.0
confluent-kafka